+++ zarrs/src/array/codec/bytes_to_bytes/blosc/blosc_codec.rs
@@ -159,6 +159,10 @@ impl CodecTraits for BloscCodec {
 
 #[cfg_attr(feature = "async", async_trait::async_trait)]
 impl BytesToBytesCodecTraits for BloscCodec {
+    fn dynamic(self: Arc<Self>) -> Arc<dyn BytesToBytesCodecTraits> {
+        self as Arc<dyn BytesToBytesCodecTraits>
+    }
+
     fn recommended_concurrency(
         &self,
         _decoded_representation: &BytesRepresentation,