+++ zarrs/src/array/codec/bytes_to_bytes/zstd/zstd_codec.rs
@@ -66,6 +66,10 @@ impl CodecTraits for ZstdCodec {
 
 #[cfg_attr(feature = "async", async_trait::async_trait)]
 impl BytesToBytesCodecTraits for ZstdCodec {
+    fn dynamic(self: Arc<Self>) -> Arc<dyn BytesToBytesCodecTraits> {
+        self as Arc<dyn BytesToBytesCodecTraits>
+    }
+
     fn recommended_concurrency(
         &self,
         _decoded_representation: &BytesRepresentation,